moved utility functions and monkey patches out of scraper.py in own files
