+++ website/frontend/management/commands/_monkeypatches.py
@@ -1,0 +1,56 @@
+# subprocess.check_output appeared in python 2.7.
+# Linerva only has 2.6
+import subprocess
+def check_output(*popenargs, **kwargs):
+    r"""Run command with arguments and return its output as a byte string.
+
+    If the exit code was non-zero it raises a CalledProcessError.  The
+    CalledProcessError object will have the return code in the returncode
+    attribute and output in the output attribute.
+
+    The arguments are the same as for the Popen constructor.  Example:
+
+    >>> check_output(["ls", "-l", "/dev/null"])
+    'crw-rw-rw- 1 root root 1, 3 Oct 18  2007 /dev/null\n'
+
+    The stdout argument is not allowed as it is used internally.
+    To capture standard error in the result, use stderr=STDOUT.
+
+    >>> check_output(["/bin/sh", "-c",
+    ...               "ls -l non_existent_file ; exit 0"],
+    ...              stderr=STDOUT)
+    'ls: non_existent_file: No such file or directory\n'
+    """
+    from subprocess import PIPE, CalledProcessError, Popen
+    if 'stdout' in kwargs:
+        raise ValueError('stdout argument not allowed, it will be overridden.')
+    process = Popen(stdout=PIPE, *popenargs, **kwargs)
+    output, unused_err = process.communicate()
+    retcode = process.poll()
+    if retcode:
+        cmd = kwargs.get("args")
+        if cmd is None:
+            cmd = popenargs[0]
+        raise CalledProcessError(retcode, cmd, output=output)
+    return output
+    
+if not hasattr(subprocess, 'check_output'):
+    subprocess.check_output = check_output
+
+# Begin hot patch for https://bugs.launchpad.net/bugs/788986
+# Ick.
+import sys
+from BeautifulSoup import BeautifulSoup
+def bs_fixed_getText(self, separator=u""):
+    bsmod = sys.modules[BeautifulSoup.__module__]
+    if not len(self.contents):
+        return u""
+    stopNode = self._lastRecursiveChild().next
+    strings = []
+    current = self.contents[0]
+    while current is not stopNode:
+        if isinstance(current, bsmod.NavigableString):
+            strings.append(current)
+        current = current.next
+    return separator.join(strings)
+sys.modules[BeautifulSoup.__module__].Tag.getText = bs_fixed_getText
